# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Product App
host: localhost:10010
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /mentors:
    x-swagger-router-controller: get_products
    get:
      tags:
      - "mentor"
      summary: Find mentor by name
      operationId: getMentors
      description: "This can only be done by the logged in user."
      produces:
      - "application/xml"
      - "application/json"
      parameters:
        - name: mentorname
          in: query
          description: 'Mentor name '
          required: true
          type: string
        - name: experience
          in: query
          description: 'An experience of a mentors'
          required: true
          type: string
        - name: course
          in: query
          description: 'Mentor course '
          required: true
          type: string
        - name: rating
          in: query
          description: 'Short list of rating,and common rating of Mentor '
          required: true
          type: string

      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Mentors"
            xml:
              name: mentors
        500:
          description: An unexpected error occured.
          schema:
            $ref: "#/definitions/Error"
  /mentors/{mentorname}:
    x-swagger-router-controller: get_products
    get:
      tags:
      - "mentor"
      summary: Find mentor by name
      operationId: getMentor
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "mentorname"
        in: "path"
        description: "The name that needs to be fetched. "
        required: true
        type: "string"
      - name: surname
        in: "query"
        description: The surname of the mentor
        type: string
      - name: nickname
        in: query
        description: The nickname of the mentor
        type: string
      - name: Experience
        in: query
        description: The Mentor Experience
        type: string
      - name:  passed people
        in: query
        description: Value of passed people 
        type: string
      - name: value
        in: query
        description: Number of people mentoring the mentor 
        type: string
      - name: rating
        in: query
        description: The rating of the mentor
        type: string
      - name: place
        in: query
        description: Place mentor in the ranking
        type: string

      responses:
        200:
          description: A product
          schema:
            $ref: "#/definitions/Mentor"
        404:
          description: The produts does not exist
        500:
          $ref: "#/responses/Standard500ErrorResponse"
    put:
      tags:
      - "mentor"
      summary: Updated mentor
      operationId: updateMentor
      produces:
      - "application/xml"
      - "application/json"
      parameters:
        - name: "mentorname"
          in: "path"
          description: "The name that needs to be fetched. "
          required: true
          type: "string"
        - name: name
          in: query
          description: The name of the mentor
          type: string
        - name: surname
          in: query
          description: The surname of the mentor
          type: string
        - name: Experience
          in: query
          description: The Mentor Experience
          type: string
        - name:  passed people
          in: query
          description: Value of passed people 
          type: string
        - name: value
          in: query
          description: Number of people mentoring the mentor 
          type: string
        - name: rating
          in: query
          description: The rating of the mentor
          type: string
        - name: place
          in: query
          description: Place mentor in the ranking
          type: string


      responses:
        200:
          description: A product
          schema:
            $ref: "#/definitions/Mentor"
        404:
          description: The produts does not exist
        500:
          $ref: "#/responses/Standard500ErrorResponse"
    post:
      tags:
      - "mentor"
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: upgradeMentor
      produces:
      - "application/xml"
      - "application/json"
      parameters:
        - name: "mentorname"
          in: "path"
          description: "The name that needs to be fetched. "
          required: true
          type: "string"
        - name: name
          in: query
          description: The name of the mentor
          type: string
        - name: surname
          in: query
          description: The surname of the mentor
          type: string
        - name: Courses
          in: query
          description: Kind of courses
          type: string
        - name: experience
          in: query
          description: The experience of the mentor
          type: string
        - name: contact
          in: query
          description: The contact
          type: string
        - name: All people
          in: query
          description: People who have a mentor
          type: string
        - name: rating
          in: query
          description: The rating of mentor
          type: string


      responses:
        200:
          description: A product
          schema:
            $ref: "#/definitions/Mentor"
        404:
          description: The produts does not exist
        500:
          $ref: "#/responses/Standard500ErrorResponse"
...
definitions:
...
  Mentors:
    type: object
    properties:
      name:
        type: string
        description: The name of the mentor
      surname:
        type: string
        description: The surname of the mentor
      nickname:
        type: string
        description: The nickname of the mentor
      courses:
        type: string
        description: Kind of business
      rating:
        type: string
        description: The rating of the course
  Mentor:
    type: object
    properties:
      name:
        type: string
        description: The name of the mentor
      surname:
        type: string
        description: The surname of the mentor
      image:
        type: string
        description: The image of the mentor
      courese:
        type: string
        description: The coures of the mentor
      experience:
        type: string
        description: experience
      chat link:
        type: string
        description: link on chat
      rating:
        type: string
        description: the rating
...
  Error:
    required:
      - message
    properties:
      message:
        type: string
responses:
  Standard500ErrorResponse:
    description: An unexpected error occured
    schema:
      $ref: "#/definitions/Error"